backend/requirements.txt
Django==5.2.6
psycopg2-binary
djangorestframework
django-cors-headers
django-filter
drf-spectacular
python-json-logger